backend_py/requirements.txt
fastapi>=0.80.0,<0.113.0
uvicorn[standard]>=0.18.0,<0.31.0
strawberry-graphql[fastapi]>=0.150.0,<0.242.0
odmantic>=0.9.0
pymongo>=4.0.0,<5.0.0
motor>=3.0.0,<4.0.0
python-dotenv>=0.20.0,<2.0.0
pydantic>=1.10.0,<2.0.0